requirements.txt
Flask==2.3.2
openai
PyGithub
google-cloud-secret-manager
google-cloud-logging
tiktoken
htmlmin
python-dotenv
